adding libmpcdec to the repository, and use it on osx and linux. add patches for MACLib for osx and linux  compatibility and use it now for osx and linux builds ( windows builds did already use it)

+++ trunk/wxMusik/MUSIKEngine/3rd-Party/MAC/MACLib/APEDecompress.cpp
@@ -446,7 +446,7 @@ int CAPEDecompress::GetInfo(APE_DECOMPRESS_FIELDS Field, int nParam1, int nParam
             char * pBuffer = (char *) nParam1;
             int nMaxBytes = nParam2;
             
-            if (sizeof(WAVE_HEADER) > nMaxBytes)
+            if (sizeof(WAVE_HEADER) > (unsigned int)nMaxBytes)
             {
                 nRetVal = -1;
             }
