+++ trunk/wxMusik/MUSIKEngine/3rd-Party/MAC/MACLib/Old/APEDecompressOld.cpp
@@ -244,7 +244,7 @@ int CAPEDecompressOld::GetInfo(APE_DECOMPRESS_FIELDS Field, int nParam1, int nPa
             char * pBuffer = (char *) nParam1;
             int nMaxBytes = nParam2;
             
-            if (sizeof(WAVE_HEADER) > nMaxBytes)
+            if (sizeof(WAVE_HEADER) > (unsigned int)nMaxBytes)
             {
                 nRetVal = -1;
             }
